requirements/development.txt
-r base.txt

# Testing
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
pytest-asyncio>=0.24,<1.0
factory-boy>=3.3,<4.0
respx>=0.22,<1.0

# Linting & Formatting
ruff>=0.8,<1.0

# Type Checking
mypy>=1.13,<2.0
django-stubs>=5.1,<6.0
djangorestframework-stubs>=3.15,<4.0
types-redis>=4.6,<5.0

# Pre-commit
pre-commit>=4.0,<5.0
commitizen>=4.1,<5.0

# Security scanning
bandit>=1.8,<2.0
safety>=3.2,<4.0

# Debug
django-debug-toolbar>=4.4,<5.0
ipython>=8.30,<9.0
